move token files
