+++ src/app/components/App.tsx
@@ -12,15 +12,12 @@ import Box from './Box';
 import { activeTabSelector } from '@/selectors';
 import PluginResizerWrapper from './PluginResizer';
 import LoadingBar from './LoadingBar';
-import { darkThemeMode, lightThemeMode } from '@/stitches.config';
-import { useFigmaTheme } from '@/hooks/useFigmaTheme';
 
 function App() {
   const activeTab = useSelector(activeTabSelector);
-  const { isDarkTheme } = useFigmaTheme();
 
   return (
-    <Box css={{ backgroundColor: '$bgDefault' }} className={isDarkTheme ? darkThemeMode : lightThemeMode}>
+    <Box>
       {activeTab !== 'loading' && <LoadingBar />}
       <PluginResizerWrapper>
         <Box
