+++ src/tokens/dark.ts
@@ -14,7 +14,6 @@ export const colors = {
   bgSubtle: '#383838',
   bgSuccess: '#476656',
   bgSwitchChecked: '#0c8ce9',
-  bgToolTip: '#1e1e1e',
   borderAccent: '#105cad',
   borderBtnDisabled: '#444444',
   borderDanger: '#864537',
@@ -22,7 +21,7 @@ export const colors = {
   borderMuted: '#444444',
   borderPro: '#8a38f5',
   borderSwitch: '#444444',
-  dangerBg: '#e03e1a',
+  dangerBgEmphasis: '#e03e1a',
   dangerBgHover: '#c4381c',
   dangerFg: '#fca397',
   dangerFgOnDark: '#E05534',
@@ -30,15 +29,12 @@ export const colors = {
   fgBtnGhost: '#ffffff',
   fgBtnPrimary: '#ffffff',
   fgBtnSecondary: '#ffffff',
-  fgDanger: '#fca397',
   fgDefault: '#ffffff',
   fgOnDarkAccent: '#A0C0EF',
   fgOnDarkDanger: '#E05534',
   fgPro: '#d1a8ff',
   fgSubtle: 'rgba(255, 255, 255, .7)',
   fgSuccess: '#79d297',
-  fgToolTip: '#ffffff',
-  fgToolTipMuted: 'rgba(255, 255, 255, 0.7)',
   focus: '#394360',
   interaction: '#0c8ce9',
   interactionDisabled: 'rgba(255, 255, 255, 0.4)',
@@ -56,10 +52,13 @@ export const colors = {
   textSubtle: 'rgba(255, 255, 255, .4)',
   contextMenuBackground: '#1e1e1e',
   contextMenuBorder: '#383838',
-  contextMenuForeground: 'rgba(255, 255, 255, 1)',
+  contextMenuForeground: '#ffffff',
   contextMenuForegroundDisabled: 'rgba(255, 255, 255, 0.4)',
   contextMenuForegroundMuted: 'rgba(255, 255, 255, 0.7)',
   contextMenuSeperator: '#383838',
+  tooltipFg: '#ffffff',
+  tooltipFgMuted: 'rgba(255, 255, 255, 0.7)',
+  tooltipBg: '#1e1e1e',
 };
 
 export const shadows = {
