+++ src/tokens/light.ts
@@ -14,7 +14,6 @@ export const colors = {
   bgSubtle: '#f5f5f5',
   bgSuccess: '#cff7d3',
   bgSwitchChecked: '#0d99ff',
-  bgToolTip: '#1e1e1e',
   borderAccent: '#bde3ff',
   borderBtnDisabled: '#e6e6e6',
   borderDanger: '#ffc7c2',
@@ -22,7 +21,7 @@ export const colors = {
   borderMuted: '#e6e6e6',
   borderPro: '#9747ff',
   borderSwitch: '#e6e6e6',
-  dangerBg: '#f24822',
+  dangerBgEmphasis: '#f24822',
   dangerBgHover: '#dc3412',
   dangerFg: '#dc3412',
   dangerFgOnDark: '#E05534',
@@ -30,15 +29,12 @@ export const colors = {
   fgBtnGhost: 'rgba(0, 0, 0, .9)',
   fgBtnPrimary: '#ffffff',
   fgBtnSecondary: 'rgba(0, 0, 0, .9)',
-  fgDanger: '#dc3412',
   fgDefault: 'rgba(0, 0, 0, .9)',
   fgOnDarkAccent: '#A0C0EF',
   fgOnDarkDanger: '#E05534',
   fgPro: '#8638e5',
   fgSubtle: 'rgba(0, 0, 0, .5)',
   fgSuccess: '#009951',
-  fgToolTip: '#ffffff',
-  fgToolTipMuted: 'rgba(255, 255, 255, 0.7)',
   focus: '#e5f4ff',
   interaction: '#0d99ff',
   interactionDisabled: 'rgba(255, 255, 255, 0.4)',
@@ -56,10 +52,13 @@ export const colors = {
   textSubtle: 'rgba(0, 0, 0, .3)',
   contextMenuBackground: '#1e1e1e',
   contextMenuBorder: '#383838',
-  contextMenuForeground: 'rgba(255, 255, 255, 1)',
+  contextMenuForeground: '#ffffff',
   contextMenuForegroundDisabled: 'rgba(255, 255, 255, 0.4)',
   contextMenuForegroundMuted: 'rgba(255, 255, 255, 0.7)',
   contextMenuSeperator: '#383838',
+  tooltipFg: '#ffffff',
+  tooltipFgMuted: 'rgba(255, 255, 255, 0.7)',
+  tooltipBg: '#1e1e1e',
 };
 
 export const shadows = {
